+++ Scripts/Security/Remove-ASR.ps1
@@ -30,7 +30,7 @@ SOFTWARE.
 
 .VERSION 0.15.1
 
-.GUID a25eb685-36f4-4bb4-b825-61cf2e737a46
+.GUID 621a598a-01bf-4c90-9b1a-2cc682fa3d64
 
 .AUTHOR metablaster zebal@protonmail.com