+++ Scripts/Security/Set-Privacy.ps1
@@ -64,6 +64,7 @@ None. Set-Privacy.ps1 does not generate any output
 .NOTES
 TODO: More Windows privacy options can be set in GPO
 TODO: A script is needed to reset privacy modification to factory defaults
+TODO: Need to exclude settings which don't apply to target computer
 
 .LINK
 https://github.com/metablaster/WindowsFirewallRuleset/blob/master/Scripts/README.md